+++ CHANGELOG.md
@@ -7,6 +7,7 @@ v5.8-SNAPSHOT
 - Dropped the already disabled Storm example
 - Fully switched to Junit 5
 - Require JDK 11 or newer to build
+- Workaround for change in Unicode CLDR(and thus Java 17): they changed the short name of "September" in Locale.UK to "Sept" which causes parse errors.
 
 v5.7
 ===